face_recognition_service_hf.py
"""
Hugging Face-based face recognition service for cloud deployment.
This version doesn't require dlib and works on Streamlit Cloud.
"""
import os
import numpy as np
from PIL import Image
import base64
import io
from dotenv import load_dotenv
from huggingface_hub import InferenceClient
from database import get_all_face_encodings

load_dotenv()

class FaceRecognitionService:
    """Service for face recognition operations using Hugging Face API"""
    
    def __init__(self):
        self.tolerance = 0.6  # Lower = more strict matching
        api_key = os.environ.get("HUGGINGFACE_API_KEY")
        if not api_key:
            raise ValueError("HUGGINGFACE_API_KEY not found in environment variables. Please set it in Streamlit Cloud secrets.")
        self.client = InferenceClient(token=api_key)
        # Using a face embedding model that works well for face recognition
        self.model = "sentence-transformers/clip-ViT-B-32"
    
    def encode_face(self, image_path):
        """
        Encode a face from an image file using Hugging Face API.
        Returns the face encoding as a list, or None if no face is found.
        """
        try:
            # Read image
            with open(image_path, 'rb') as f:
                image_data = f.read()
            
            # Use image-to-text or image feature extraction
            # For face recognition, we'll use a vision model that can extract features
            # Note: This is a simplified approach. For production, consider using
            # a dedicated face recognition model from Hugging Face
            
            # Convert image to base64 for API
            image_base64 = base64.b64encode(image_data).decode('utf-8')
            
            # Use feature extraction - this will give us a vector representation
            # We'll use a vision transformer model
            try:
                # Try using image feature extraction
                # Note: Hugging Face inference API may have different endpoints
                # This is a placeholder - you may need to adjust based on available models
                
                # Alternative: Use a face detection/recognition model
                # For now, we'll use a simple approach with image embeddings
                from transformers import CLIPProcessor, CLIPModel
                import torch
                
                # Load model (this will be cached)
                if not hasattr(self, '_clip_model'):
                    self._clip_model = CLIPModel.from_pretrained("openai/clip-vit-base-patch32")
                    self._clip_processor = CLIPProcessor.from_pretrained("openai/clip-vit-base-patch32")
                
                # Process image
                image = Image.open(image_path).convert('RGB')
                inputs = self._clip_processor(images=image, return_tensors="pt")
                
                with torch.no_grad():
                    image_features = self._clip_model.get_image_features(**inputs)
                
                # Convert to numpy array and normalize
                encoding = image_features[0].numpy()
                encoding = encoding / np.linalg.norm(encoding)  # Normalize
                
                return encoding.tolist()
                
            except Exception as e:
                print(f"Error with CLIP model: {e}")
                # Fallback: Use a simpler approach
                # For now, return a basic encoding based on image features
                # This is a simplified version - in production, use a proper face recognition model
                return self._simple_image_encoding(image_path)
                
        except Exception as e:
            print(f"Error encoding face: {e}")
            return None
    
    def _simple_image_encoding(self, image_path):
        """
        Simple fallback encoding method.
        In production, replace this with a proper face recognition API call.
        """
        try:
            # Load and resize image
            img = Image.open(image_path).convert('RGB')
            img = img.resize((224, 224))
            
            # Convert to array and flatten
            img_array = np.array(img).flatten()
            
            # Normalize
            img_array = img_array / 255.0
            
            # Use PCA-like reduction to get a reasonable encoding size
            # This is a simplified approach - not ideal for face recognition
            # but works as a fallback
            encoding = img_array[:512]  # Take first 512 values
            
            # Normalize the encoding
            encoding = encoding / (np.linalg.norm(encoding) + 1e-8)
            
            return encoding.tolist()
        except:
            return None
    
    def find_matches(self, query_encoding, threshold=0.6):
        """
        Find matches for a query face encoding in the database.
        
        Args:
            query_encoding: list or numpy array of the face encoding to search for
            threshold: distance threshold for matching (lower = more strict)
        
        Returns:
            List of matches with person_id, distance, and confidence
        """
        matches = []
        
        # Get all encodings from database
        stored_encodings = get_all_face_encodings()
        
        if not stored_encodings:
            return matches
        
        # Convert query encoding to numpy array if needed
        if isinstance(query_encoding, list):
            query_encoding = np.array(query_encoding)
        
        for stored in stored_encodings:
            # Convert stored encoding to numpy array
            stored_encoding = np.array(stored['encoding'])
            
            # Calculate cosine similarity (better for normalized vectors)
            # Cosine similarity = 1 - cosine distance
            dot_product = np.dot(query_encoding, stored_encoding)
            norm_product = np.linalg.norm(query_encoding) * np.linalg.norm(stored_encoding)
            
            if norm_product == 0:
                continue
                
            cosine_similarity = dot_product / norm_product
            cosine_distance = 1 - cosine_similarity
            
            # Check if distance is below threshold
            if cosine_distance <= threshold:
                # Convert distance to confidence (0-1 scale, inverted)
                confidence = 1 - (cosine_distance / threshold)
                confidence = max(0, min(1, confidence))  # Clamp between 0 and 1
                
                matches.append({
                    'person_id': stored['person_id'],
                    'distance': float(cosine_distance),
                    'confidence': float(confidence)
                })
        
        # Sort by distance (closest first)
        matches.sort(key=lambda x: x['distance'])
        
        return matches
    
    def compare_faces(self, encoding1, encoding2, threshold=0.6):
        """
        Compare two face encodings.
        Returns True if faces match, False otherwise.
        """
        if isinstance(encoding1, list):
            encoding1 = np.array(encoding1)
        if isinstance(encoding2, list):
            encoding2 = np.array(encoding2)
        
        # Calculate cosine distance
        dot_product = np.dot(encoding1, encoding2)
        norm_product = np.linalg.norm(encoding1) * np.linalg.norm(encoding2)
        
        if norm_product == 0:
            return False, 1.0
        
        cosine_similarity = dot_product / norm_product
        cosine_distance = 1 - cosine_similarity
        
        return cosine_distance <= threshold, float(cosine_distance)

